smsmic/requirements.txt
Flask==2.2.5
requests==2.31.0
python-dotenv==1.0.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.2
Flask-Migrate==4.0.4